Reach Main Menu
🔖 This commit does the bare minimum to reach the title screen.

An unknown mod is manually invoking MixinExtras, which clashes with the module pushed by Debugify. Since I do not know the other mod, Debugify is put on the chopping block for now.

⚠️ Though this commit is able to launch to the main menu, it is still unusable due to a failure to generate worlds for an unknown reason. Expect a fix in future commits.

---

- Update, disable mod Debugify

- Add mod Crash Utilities
- Add mod Not Enough Crashes
- Add mod Cyanide

(Since this is the first successful run, the game generates some config files from mods that haven't worked before.)