Untrack/uncache per-user gender configs
It's kind of arbitrary. On clientside, you can edit your values and the values for other players as you and you only see them.

On normalside/serverside, you will see what *other people* see for *themselves*. So, if for instance, you joined my game, this file would automatically populate with my settings, and thus I have no reason to track this, especially not in the repo